Lambda client refactor - Skills
Elevates lambda connection to a single common file, and in doing so, removes need for internal connections and testing.
