+++ src/skill/createSkillFn/createSkill.test.ts
@@ -19,14 +19,6 @@ describe('Create the Skill Handler', () => {
         expect(res.statusCode).toEqual(400);
     })
 
-    it('should fail with 500, from a database connection error', async () => {
-        (client.connect as jest.Mock).mockImplementationOnce(() => {
-            throw "error"
-        })
-        const res = await handler({ body: JSON.stringify(skill) } as APIGatewayProxyEvent)
-        expect(res.statusCode).toEqual(500);
-    })
-
     it('should fail with 400, from a database query error', async () => {
         (client.query as jest.Mock).mockImplementationOnce(() => {
             throw "error"
