+++ src/skill/deleteSkillFn/deleteSkill.test.ts
@@ -20,14 +20,6 @@ describe('Delete a Skill Handler', () => {
         expect(res.statusCode).toEqual(400);
     })
 
-    it('should fail with 500, from a database connection error', async () => {
-        (client.connect as jest.Mock).mockImplementationOnce(() => {
-            throw "error"
-        })
-        const res = await handler(input as APIGatewayProxyEvent)
-        expect(res.statusCode).toEqual(500);
-    })
-
     it('should fail with 400, from a database query error', async () => {
         (client.query as jest.Mock).mockImplementationOnce(() => {
             throw "error"
