+++ src/skill/getAllSkillsFn/getAllSkills.test.ts
@@ -12,14 +12,6 @@ describe('Get All Skills Handler', () => {
         expect(res.statusCode).toEqual(200);
     })
 
-    it('should fail with 500, from a database connection error', async () => {
-        (client.connect as jest.Mock).mockImplementationOnce(() => {
-            throw "error"
-        })
-        const res = await handler({} as APIGatewayProxyEvent)
-        expect(res.statusCode).toEqual(500);
-    })
-
     it('should fail with 400, from a database query error', async () => {
         (client.query as jest.Mock).mockImplementationOnce(() => {
             throw "error"
